First functional version of proccon and procpeering
Proccon accepts HTTP POST JSON messages with metrics.
It create Server and Software nodes as needed and add network info to
Software nodes.
It decides where to put each network info based on the cmdline of the
process.
If it does not find any suitable Software, it creates an "others"
Software, with this approach we can capture most of the connections
between nodes, even though we haven't the specific Software.
Clean connections with a "garbage collector", delete connections not
updated more than a configurable time. It runs periodically
(configurable).

Procpeering create edges between Software nodes that match (one have
a connetion to an endpoint, the other have the listener for that
endpoint).
Creation of edges is always initilizated by client connections (no by
process listeners)
It handles also the delete of edges if the match is no longer available.