proccon probe: allow a prefix to distinguish different priv networks
If we have nodes sharing same private subnets, the previous
implementation does not provide any option to differentiate those
networks, so procpeering could end linking nodes incorrectly.

This new feature could prefix all IPs with a value in the metric, so we
move the responsability of knowing how subnets are used outside of
skydive.