Use metric timestamp as CreatedAt/UpdatedAt value
Until now, the CreatedAt/UpdatedAt values stored with each connection
were set once received by Skydive, using time.Now().
This could generate incorrect data when the communication between
clients and Skydive is unavailable, and clients start caching values.
When the connection is restored, all those cached metrics are inserted
with the same value in Skydive, losing that piece of information.

Using the metric timestamp could also generate some problems if the
clocks of skydive server and clients are not synced.
Metrics from a client with its clock in the past could be deleted by the
garbage collector, but the threshold to delete those metric will be
usually much larger than the clock sync difference.